Reorganize discogs files
