+++ manifest.json
@@ -64,9 +64,9 @@
         "src/bandcamp/pages/page-music.js",
         "src/bandcamp/styles.css",
         "src/config.js",
-        "src/discogs/discogs-csv.js",
-        "src/discogs/discogs.js",
-        "src/discogs/genres.js",
+        "src/discogs/classes/discogs-csv.js",
+        "src/discogs/modules/discogs.js",
+        "src/discogs/modules/genres.js",
         "src/modules/chrome.js",
         "src/modules/html.js",
         "src/modules/key-generator.js",
