+++ src/bandcamp/modules/mapping.js
@@ -1,4 +1,4 @@
-import { getGenreByStyle } from "../../discogs/genres.js";
+import { getGenreByStyle } from "../../discogs/modules/genres.js";
 import { isEmptyObject, isObject, isString } from "../../modules/utils.js";
 
 let mapping = {};
