+++ src/discogs/modules/genres.js
