+++ src/popup/popup.js
@@ -1,6 +1,6 @@
 import { Release } from "../app/release.js";
 import { getCurrentTab, getExtensionManifest } from "../modules/chrome.js";
-import { loadDiscogsGenres } from "../discogs/genres.js";
+import { loadDiscogsGenres } from "../discogs/modules/genres.js";
 import { loadKeywordMapping } from "../bandcamp/modules/mapping.js";
 import config from "../config.js";
 import { setupStorageTab } from "./tabs/storage_tab.js";
